[TICTIC-1122] Add long press floating menu.
Create menu by menu resource, compatible with Android's floating context
menu.

Now, we only support flat menu (no sub-menu) with basic data.

For 1 ~ 3 items, we use a horizontal linear layout to show items.
For 4 items, we use a two line layout.
For items >= 5, we use a list with a close button on bottom.

Style of the floating context menu keep same as Ticwear 1's Action
Layout. That is no swipe, fade in/out animation with a blur behind.

Change-Id: I17dd0a491e2fcef4d9f9775cae041756832b7a30